Feature/cashback interface

+++ backend/src/main/resources/db/migration/V19__create_clinic_configuration_table.sql
@@ -4,7 +4,7 @@
 -- =====================================================
 
 CREATE TABLE clinic_configuration (
-    id TINYINT NOT NULL,
+    id INT NOT NULL,
     clinic_name VARCHAR(150) NOT NULL,
     address VARCHAR(500) NULL,
     phone VARCHAR(30) NULL,
